Add camera and image upload functionality
When a property form is filled out, it will take the user to an add photo page. The user will have to take a picture. It will show the picture on the screen once taken. Then the user clicks upload and it will upload to the server and return them to the listing page